ingestion/requirements.txt
bcrypt==4.1.2
docx2txt==0.8
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-mongodb==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
llama-index==0.11.11
llama-index-embeddings-openai==0.2.5
llama-index-vector-stores-mongodb==0.3.0
openai==1.43.0
pymongo==4.9.2
pypdf==4.1.0
python-dotenv==1.0.1
streamlit==1.30.0
tiktoken==0.7.0
